+++ src/ByteDev.Time/TimeMax.cs
@@ -1,0 +1,13 @@
+﻿namespace ByteDev.Time.UnitTests
+{
+    public static class TimeMax
+    {
+        public static readonly int Hours = 23;
+
+        public static readonly int Minutes = 59;
+
+        public static readonly int Seconds = 59;
+
+        public static readonly int Milliseconds = 999;
+    }
+}