+++ tests/ByteDev.Time.UnitTests/DateTimeExtensionsTests.cs
@@ -137,5 +137,159 @@ namespace ByteDev.Time.UnitTests
                 Assert.That(result, Is.EqualTo(expected));
             }
         }
+
+        [TestFixture]
+        public class WithHour
+        {
+            [Test]
+            public void WhenCalled_ThenReturnsWithNewHour()
+            {
+                var expected = new DateTime(2020, 1, 10, 10, 59, 30);
+
+                var sut = new DateTime(2020, 1, 10, 23, 59, 30);
+
+                var result = sut.WithHour(expected.Hour);
+
+                Assert.That(result, Is.EqualTo(expected));
+            }
+        }
+
+        [TestFixture]
+        public class WithMinute
+        {
+            [Test]
+            public void WhenCalled_ThenReturnsWithNewMinute()
+            {
+                var expected = new DateTime(2020, 1, 10, 10, 59, 30);
+
+                var sut = new DateTime(2020, 1, 10, 10, 10, 30);
+
+                var result = sut.WithMinute(expected.Minute);
+
+                Assert.That(result, Is.EqualTo(expected));
+            }
+        }
+
+        [TestFixture]
+        public class WithSecond
+        {
+            [Test]
+            public void WhenCalled_ThenReturnsWithNewSecond()
+            {
+                var expected = new DateTime(2020, 1, 10, 20, 30, 40);
+
+                var sut = new DateTime(2020, 1, 10, 20, 30, 30);
+
+                var result = sut.WithSecond(expected.Second);
+
+                Assert.That(result, Is.EqualTo(expected));
+            }
+        }
+
+        [TestFixture]
+        public class WithMillisecond
+        {
+            [Test]
+            public void WhenCalled_ThenReturnsWithNewMillisecond()
+            {
+                var expected = new DateTime(2020, 1, 10, 20, 30, 40, 500);
+
+                var sut = new DateTime(2020, 1, 10, 20, 30, 40, 999);
+
+                var result = sut.WithMillisecond(expected.Millisecond);
+
+                Assert.That(result, Is.EqualTo(expected));
+            }
+        }
+
+        [TestFixture]
+        public class EndOfDay
+        {
+            [Test]
+            public void WhenCalled_ThenReturnsEndOfDay()
+            {
+                var sut = new DateTime(2020, 6, 13, 10, 20, 30, 500);
+                
+                var result = sut.EndOfDay();
+                
+                Assert.That(result, Is.EqualTo(new DateTime(2020, 6, 13, TimeMax.Hours, TimeMax.Minutes, TimeMax.Seconds, TimeMax.Milliseconds)));
+            }
+        }
+
+        [TestFixture]
+        public class EndOfMonth
+        {
+            [Test]
+            public void WhenCalled_ThenReturnsEndOfMonth()
+            {
+                var sut = new DateTime(2020, 6, 13, 10, 20, 30, 500);
+
+                var result = sut.EndOfMonth();
+                
+                Assert.That(result, Is.EqualTo(new DateTime(2020, 6, 30, TimeMax.Hours, TimeMax.Minutes, TimeMax.Seconds, TimeMax.Milliseconds)));
+            }
+        }
+
+        [TestFixture]
+        public class EndOfYear
+        {
+            [Test]
+            public void WhenCalled_ThenReturnsEndOfDay()
+            {
+                var sut = new DateTime(2020, 6, 13, 10, 20, 30, 500);
+                
+                var result = sut.EndOfYear();
+                
+                Assert.That(result, Is.EqualTo(new DateTime(2020, 12, 31, TimeMax.Hours, TimeMax.Minutes, TimeMax.Seconds, TimeMax.Milliseconds)));
+            }
+        }
+
+        [TestFixture]
+        public class EqualsDate
+        {
+            [Test]
+            public void WhenDateIsEqual_ThenReturnTrue()
+            {
+                var sut = new DateTime(2020, 6, 13, 10, 20, 30, 500);
+
+                var result = sut.EqualsDate(new DateTime(sut.Year, sut.Month, sut.Day, sut.Hour + 1, sut.Minute + 1, sut.Second + 1, sut.Millisecond + 1));
+
+                Assert.That(result, Is.True);
+            }
+
+            [Test]
+            public void WhenDateIsNotEqual_ThenReturnFalse()
+            {
+                var sut = new DateTime(2020, 6, 13, 10, 20, 30, 500);
+
+                var result = sut.EqualsDate(sut.WithDay(sut.Day + 1));
+
+                Assert.That(result, Is.False);
+            }
+        }
+
+        [TestFixture]
+        public class EqualsTime
+        {
+            [Test]
+            public void WhenTimeIsEqual_ThenReturnTrue()
+            {
+                var sut = new DateTime(2020, 6, 13, 10, 20, 30, 500);
+
+                var result = sut.EqualsTime(new DateTime(sut.Year + 1, sut.Month + 1, sut.Day + 1, sut.Hour, sut.Minute, sut.Second, sut.Millisecond));
+
+                Assert.That(result, Is.True);
+            }
+
+            [Test]
+            public void WhenTimeIsNotEqual_ThenReturnFalse()
+            {
+                var sut = new DateTime(2020, 6, 13, 10, 20, 30, 500);
+
+                var result = sut.EqualsTime(sut.WithMillisecond(sut.Millisecond + 1));
+
+                Assert.That(result, Is.False);
+            }
+        }
     }
 }