+++ docs/RELEASE-NOTES.md
@@ -1,5 +1,25 @@
 # Release Notes
 
+## 1.3.0 - 15 July 2020
+
+Breaking changes:
+- (None)
+
+New features:
+- Added DateTimeExtensions.EndOfDay
+- Added DateTimeExtensions.EndOfMonth
+- Added DateTimeExtensions.EndOfYear
+- Added DateTimeExtensions.EqualsDate
+- Added DateTimeExtensions.EqualsTime
+- Added DateTimeExtensions.WithHour
+- Added DateTimeExtensions.WithMinute
+- Added DateTimeExtensions.WithSecond
+- Added DateTimeExtensions.WithMillisecond
+- Added TimeMax
+
+Bug fixes / internal changes:
+- (None)
+
 ## 1.2.0 - 17 April 2020
 
 Breaking changes: